# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Filament DB API",
    "description": "REST API for managing 3D printing filament profiles, nozzles, and printers. Supports filament inheritance (parent/variant), per-printer per-nozzle calibrations, INI profile export/import, and OpenPrintTag NFC binary generation.",
    "version": "1.8.4",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/api/nozzles": {
      "get": {
        "tags": ["Nozzles"],
        "summary": "List nozzles",
        "description": "Returns all non-deleted nozzles, optionally filtered by diameter, type, or high-flow flag.",
        "parameters": [
          {
            "name": "diameter",
            "in": "query",
            "schema": { "type": "number" },
            "description": "Filter by nozzle diameter in mm"
          },
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by nozzle type (e.g. Brass, Hardened Steel)"
          },
          {
            "name": "highFlow",
            "in": "query",
            "schema": { "type": "string", "enum": ["true", "false"] },
            "description": "Filter by high-flow capability"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of nozzles sorted by diameter then type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Nozzle" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Nozzles"],
        "summary": "Create nozzle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NozzleInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created nozzle",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Nozzle" }
              }
            }
          }
        }
      }
    },
    "/api/nozzles/{id}": {
      "get": {
        "tags": ["Nozzles"],
        "summary": "Get nozzle by ID",
        "parameters": [
          { "$ref": "#/components/parameters/NozzleId" }
        ],
        "responses": {
          "200": {
            "description": "Nozzle object",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Nozzle" }
              }
            }
          },
          "404": {
            "description": "Nozzle not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Nozzles"],
        "summary": "Update nozzle",
        "parameters": [
          { "$ref": "#/components/parameters/NozzleId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NozzleInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated nozzle",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Nozzle" }
              }
            }
          },
          "404": {
            "description": "Nozzle not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Nozzles"],
        "summary": "Delete nozzle",
        "description": "Soft-deletes a nozzle. Blocked if any filament references it in compatibleNozzles or calibrations.",
        "parameters": [
          { "$ref": "#/components/parameters/NozzleId" }
        ],
        "responses": {
          "200": {
            "description": "Deletion confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Deleted" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete — referenced by filaments",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Nozzle not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/printers/{id}": {
      "get": {
        "tags": ["Printers"],
        "summary": "Get printer by ID",
        "parameters": [
          { "$ref": "#/components/parameters/PrinterId" }
        ],
        "responses": {
          "200": {
            "description": "Printer with populated nozzles",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Printer" }
              }
            }
          },
          "404": {
            "description": "Printer not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Printers"],
        "summary": "Update printer",
        "parameters": [
          { "$ref": "#/components/parameters/PrinterId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PrinterInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated printer",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Printer" }
              }
            }
          },
          "404": {
            "description": "Printer not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Printers"],
        "summary": "Delete printer",
        "description": "Soft-deletes a printer. Blocked if any filament calibration references it.",
        "parameters": [
          { "$ref": "#/components/parameters/PrinterId" }
        ],
        "responses": {
          "200": {
            "description": "Deletion confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Deleted" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete — referenced by filament calibrations",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Printer not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/openprinttag/import": {
      "post": {
        "tags": ["OpenPrintTag"],
        "summary": "Import OpenPrintTag materials",
        "description": "Import selected OpenPrintTag materials into Filament DB. Materials are upserted by name (not name+vendor). If a filament with the same name exists under a different vendor, the import is skipped with an error.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["slugs"],
                "properties": { "slugs": { "type": "array", "items": { "type": "string" } } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Import result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportResult" } } } },
          "400": { "description": "Missing slugs or database not loaded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "NozzleId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "MongoDB ObjectId of the nozzle"
      },
      "PrinterId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "MongoDB ObjectId of the printer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
...
      "Nozzle": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string", "example": "0.4mm Brass" },
          "diameter": { "type": "number", "example": 0.4 },
          "type": { "type": "string", "example": "Brass" },
          "highFlow": { "type": "boolean", "default": false },
          "hardened": { "type": "boolean", "default": false },
          "notes": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["name", "diameter", "type"]
      },
      "NozzleInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "diameter": { "type": "number" },
          "type": { "type": "string" },
          "highFlow": { "type": "boolean" },
          "hardened": { "type": "boolean" },
          "notes": { "type": "string" }
        },
        "required": ["name", "diameter", "type"]
      },
      "Printer": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string", "example": "Prusa Core One" },
          "manufacturer": { "type": "string", "example": "Prusa" },
          "printerModel": { "type": "string", "example": "Core One" },
          "installedNozzles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Nozzle" },
            "description": "Populated nozzle objects"
          },
          "notes": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["name", "manufacturer", "printerModel"]
      },
      "PrinterInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "manufacturer": { "type": "string" },
          "printerModel": { "type": "string" },
          "installedNozzles": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Array of Nozzle ObjectIds"
          },
          "notes": { "type": "string" }
        },
        "required": ["name", "manufacturer", "printerModel"]
      },
      "ImportResult": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "total": { "type": "integer" },
          "created": { "type": "integer" },
          "updated": { "type": "integer" },
          "skipped": { "type": "integer" }
        }
      },
...
    }
  }
}